src/requirements.txt
Flask
requests
mysql-connector-python
numpy
matplotlib
mysqlclient
SQLAlchemy